requirements.txt
asttokens>=2.0.5
async-generator>=1.10
attrs>=21.4.0
backcall>=0.2.0
beautifulsoup4>=4.11.1
branca>=0.5.0
Brotli>=1.0.9
brotlipy>=0.7.0
certifi>=2022.6.15
cffi>=1.15.1
charset-normalizer>=2.1.0
click>=8.1.3
click-plugins>=1.1.1
cligj>=0.7.2
colorama>=0.4.5
cryptography>=37.0.4
cycler>=0.11.0
dash>=2.6.1
dash-bootstrap-components>=1.2.1
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
debugpy>=1.5.1
decorator>=5.1.1
descartes>=1.1.0
entrypoints>=0.4
et-xmlfile>=1.1.0
executing>=0.8.3
fastjsonschema>=2.16.2
Flask>=2.2.2
Flask-Compress>=1.12
folium>=0.12.1
fonttools>=4.34.4
fuzzywuzzy>=0.18.0
googlemaps>=4.6.0
h11>=0.13.0
haversine>=2.6.0
idna>=3.3
importlib-metadata>=4.12.0
ipykernel>=6.9.1
ipython>=8.3.0
itsdangerous>=2.1.2
jedi>=0.18.1
Jinja2>=3.1.2
joblib>=1.1.0
jupyter-client>=7.2.2
jupyter-core>=4.10.0
kiwisolver>=1.4.4
lxml>=4.8.0
mapclassify>=2.4.3
MarkupSafe>=2.1.1
matplotlib>=3.5.2
matplotlib-inline>=0.1.2
munch>=2.5.0
nbformat>=5.5.0
geopandas>=0.13.2
fiona>=1.9.4
seaborn>=0.12.0
shapely>=1.8.2
plotly>=5.10.0
plotly-express>=0.4.1
nest-asyncio>=1.5.5
networkx>=2.8.5
numpy>=1.23.1
openpyxl>=3.0.10
outcome>=1.2.0
packaging>=21.3
pandas>=1.4.3
parso>=0.8.3
Pillow>=9.2.0
plotly>=5.10.0
prompt-toolkit>=3.0.20
pure-eval>=0.2.2
pycparser>=2.21
Pygments>=2.11.2
pyOpenSSL>=22.0.0
pyparsing>=3.0.9
PySocks>=1.7.1
python-dateutil>=2.8.2
python-dotenv>=0.21.0
python-Levenshtein>=0.12.2
pytz>=2022.1
pywin32>=311
pyzmq>=22.3.0
requests>=2.28.1
Rtree>=1.0.0
Shapely>=1.8.2
scikit-learn>=1.1.1
scipy>=1.8.1
selenium>=4.4.3
six>=1.16.0
sniffio>=1.3.0
sortedcontainers>=2.4.0
soupsieve>=2.3.2
tenacity>=8.0.1
threadpoolctl>=3.1.0
tornado>=6.1
tqdm>=4.64.0
traitlets>=5.1.1
trio>=0.21.0
trio-websocket>=0.9.2
unicodedata2>=14.0.0
urllib3>=1.26.11
utm>=0.7.0
wcwidth>=0.2.5
Werkzeug>=2.2.2
wheel>=0.37.1
win-inet-pton>=1.1.0
wincertstore>=0.2
wsproto>=1.2.0
xyzservices>=2022.6.0
zipp>=3.8.1
